Database restore: restore custom columns from field_metadata in saved preferences.

+++ src/calibre/library/prefs.py
@@ -80,13 +80,16 @@ class DBPrefs(dict):
             import traceback
             traceback.print_exc()
 
-    def read_serialized(self, library_path):
+    @classmethod
+    def read_serialized(self, library_path, recreate_prefs=False):
         try:
             from_filename = os.path.join(library_path,
                     'metadata_db_prefs_backup.json')
             with open(from_filename, "rb") as f:
-                self.clear()
                 d = json.load(f, object_hook=from_json)
+                if not recreate_prefs:
+                    return d
+                self.clear()
                 self.db.conn.execute('DELETE FROM preferences')
                 for k,v in d.iteritems():
                     raw = self.to_raw(v)
@@ -95,6 +98,9 @@ class DBPrefs(dict):
                 self.db.conn.commit()
                 self.clear()
                 self.update(d)
+                return d
         except:
             import traceback
             traceback.print_exc()
+            raise
+        return None
